requirements.txt
absl-py==1.4.0 ; python_version >= "3.8" and python_version < "3.9"
ale-py==0.7.5 ; python_version >= "3.8" and python_version < "3.9"
appdirs==1.4.4 ; python_version >= "3.8" and python_version < "3.9"
atomicwrites==1.4.1 ; python_version >= "3.8" and python_version < "3.9" and sys_platform == "win32"
attrs==23.1.0 ; python_version >= "3.8" and python_version < "3.9"
beartype==0.14.1 ; python_version >= "3.8" and python_version < "3.9"
box2d-py==2.3.5 ; python_version >= "3.8" and python_version < "3.9"
certifi==2023.7.22 ; python_version >= "3.8" and python_version < "3.9"
cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.9"
charset-normalizer==3.2.0 ; python_version >= "3.8" and python_version < "3.9"
click==8.1.6 ; python_version >= "3.8" and python_version < "3.9"
cloudpickle==2.2.1 ; python_version >= "3.8" and python_version < "3.9"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.9" and (sys_platform == "win32" or platform_system == "Windows")
contourpy==1.1.0 ; python_version >= "3.8" and python_version < "3.9"
cvxpy==1.3.2 ; python_version >= "3.8" and python_version < "3.9"
cycler==0.11.0 ; python_version >= "3.8" and python_version < "3.9"
cython==3.0.0 ; python_version >= "3.8" and python_version < "3.9"
d4rl @ git+https://github.com/ezhang7423/d4rl-installable.git@HEAD ; python_version >= "3.8" and python_version < "3.9"
dm-control==1.0.14 ; python_version >= "3.8" and python_version < "3.9"
dm-env==1.6 ; python_version >= "3.8" and python_version < "3.9"
dm-tree==0.1.8 ; python_version >= "3.8" and python_version < "3.9"
docker-pycreds==0.4.0 ; python_version >= "3.8" and python_version < "3.9"
ecos==2.0.12 ; python_version >= "3.8" and python_version < "3.9"
einops==0.6.1 ; python_version >= "3.8" and python_version < "3.9"
executing==1.2.0 ; python_version >= "3.8" and python_version < "3.9"
ezjaxtyping==0.2.20 ; python_version >= "3.8" and python_version < "3.9"
eztils[torch]==0.4.73 ; python_version >= "3.8" and python_version < "3.9"
filelock==3.12.2 ; python_version >= "3.8" and python_version < "3.9"
fonttools==4.41.1 ; python_version >= "3.8" and python_version < "3.9"
gitdb==4.0.10 ; python_version >= "3.8" and python_version < "3.9"
gitpython==3.1.32 ; python_version >= "3.8" and python_version < "3.9"
glfw==2.6.2 ; python_version >= "3.8" and python_version < "3.9"
gtimer==1.0.0b5 ; python_version >= "3.8" and python_version < "3.9"
gym-notices==0.0.8 ; python_version >= "3.8" and python_version < "3.9"
gym==0.23.1 ; python_version >= "3.8" and python_version < "3.9"
gym[all]==0.23.1 ; python_version >= "3.8" and python_version < "3.9"
h5py==3.9.0 ; python_version >= "3.8" and python_version < "3.9"
idna==3.4 ; python_version >= "3.8" and python_version < "3.9"
imageio==2.31.1 ; python_version >= "3.8" and python_version < "3.9"
importlib-metadata==6.8.0 ; python_version >= "3.8" and python_version < "3.9"
importlib-resources==6.0.0 ; python_version >= "3.8" and python_version < "3.9"
iniconfig==2.0.0 ; python_version >= "3.8" and python_version < "3.9"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "3.9"
kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.9"
labmaze==1.0.6 ; python_version >= "3.8" and python_version < "3.9"
lockfile==0.12.2 ; python_version >= "3.8" and python_version < "3.9"
loguru==0.7.0 ; python_version >= "3.8" and python_version < "3.9"
lxml==4.9.3 ; python_version >= "3.8" and python_version < "3.9"
lz4==4.3.2 ; python_version >= "3.8" and python_version < "3.9"
markdown-it-py==3.0.0 ; python_version >= "3.8" and python_version < "3.9"
markupsafe==2.1.3 ; python_version >= "3.8" and python_version < "3.9"
matplotlib==3.7.2 ; python_version >= "3.8" and python_version < "3.9"
mdurl==0.1.2 ; python_version >= "3.8" and python_version < "3.9"
mjrl @ git+https://github.com/aravindr93/mjrl@master ; python_version >= "3.8" and python_version < "3.9"
mpmath==1.3.0 ; python_version >= "3.8" and python_version < "3.9"
mujoco-py==1.50.1.68 ; python_version >= "3.8" and python_version < "3.9"
mujoco==2.3.7 ; python_version >= "3.8" and python_version < "3.9"
networkx==3.1 ; python_version >= "3.8" and python_version < "3.9"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.9"
opencv-python==4.8.0.74 ; python_version >= "3.8" and python_version < "3.9"
osqp==0.6.3 ; python_version >= "3.8" and python_version < "3.9"
packaging==21.3 ; python_version >= "3.8" and python_version < "3.9"
patchelf==0.17.2.1 ; python_version >= "3.8" and python_version < "3.9"
pathtools==0.1.2 ; python_version >= "3.8" and python_version < "3.9"
pillow==10.0.0 ; python_version >= "3.8" and python_version < "3.9"
pluggy==1.2.0 ; python_version >= "3.8" and python_version < "3.9"
portalocker==2.7.0 ; python_version >= "3.8" and python_version < "3.9"
protobuf==4.23.4 ; python_version >= "3.8" and python_version < "3.9"
psutil==5.9.5 ; python_version >= "3.8" and python_version < "3.9"
py==1.11.0 ; python_version >= "3.8" and python_version < "3.9"
pybullet==3.2.5 ; python_version >= "3.8" and python_version < "3.9"
pycparser==2.21 ; python_version >= "3.8" and python_version < "3.9"
pydantic==1.10.12 ; python_version >= "3.8" and python_version < "3.9"
pygame==2.1.0 ; python_version >= "3.8" and python_version < "3.9"
pygments==2.15.1 ; python_version >= "3.8" and python_version < "3.9"
pyopengl==3.1.7 ; python_version >= "3.8" and python_version < "3.9"
pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.9"
pytest==7.0.1 ; python_version >= "3.8" and python_version < "3.9"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.9"
pywin32==306 ; python_version >= "3.8" and python_version < "3.9" and platform_system == "Windows"
pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "3.9"
qdldl==0.1.7.post0 ; python_version >= "3.8" and python_version < "3.9"
requests==2.31.0 ; python_version >= "3.8" and python_version < "3.9"
rich==13.4.2 ; python_version >= "3.8" and python_version < "3.9"
scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.9"
scs==3.2.3 ; python_version >= "3.8" and python_version < "3.9"
sentry-sdk==1.28.1 ; python_version >= "3.8" and python_version < "3.9"
setproctitle==1.3.2 ; python_version >= "3.8" and python_version < "3.9"
setuptools==68.0.0 ; python_version >= "3.8" and python_version < "3.9"
shellingham==1.5.0.post1 ; python_version >= "3.8" and python_version < "3.9"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.9"
smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.9"
swig==4.1.1 ; python_version >= "3.8" and python_version < "3.9"
sympy==1.12 ; python_version >= "3.8" and python_version < "3.9"
tensorboardx==2.6.1 ; python_version >= "3.8" and python_version < "3.9"
termcolor==2.3.0 ; python_version >= "3.8" and python_version < "3.9"
tomli==2.0.1 ; python_version >= "3.8" and python_version < "3.9"
torch==2.0.1 ; python_version >= "3.8" and python_version < "3.9"
torchtyping==0.1.4 ; python_version >= "3.8" and python_version < "3.9"
torchvision==0.15.2 ; python_version >= "3.8" and python_version < "3.9"
tqdm==4.65.0 ; python_version >= "3.8" and python_version < "3.9"
typeguard==4.0.0 ; python_version >= "3.8" and python_version < "3.9"
typer[all]==0.9.0 ; python_version >= "3.8" and python_version < "3.9"
types-python-dateutil==2.8.19.14 ; python_version >= "3.8" and python_version < "3.9"
types-setuptools==68.0.0.3 ; python_version >= "3.8" and python_version < "3.9"
typing-extensions==4.7.1 ; python_version >= "3.8" and python_version < "3.9"
urllib3==2.0.4 ; python_version >= "3.8" and python_version < "3.9"
varname==0.11.2 ; python_version >= "3.8" and python_version < "3.9"
wandb==0.15.7 ; python_version >= "3.8" and python_version < "3.9"
win32-setctime==1.1.0 ; python_version >= "3.8" and python_version < "3.9" and sys_platform == "win32"
zipp==3.16.2 ; python_version >= "3.8" and python_version < "3.9"
